+++ idea/testData/quickfix/createLabel/breakInOuterLoop.kt
@@ -1,0 +1,9 @@
+// "Create label foo@" "true"
+
+fun test() {
+    while (true) {
+        while (true) {
+            break@<caret>foo
+        }
+    }
+}