+++ idea/testData/quickfix/createLabel/continueNoLoop.kt
@@ -1,0 +1,8 @@
+// "Create label foo@" "false"
+// ACTION: Convert to expression body
+// ERROR: The label '@foo' does not denote a loop
+// ERROR: Unresolved reference: @foo
+
+fun test() {
+    continue@<caret>foo
+}