+++ idea/testData/quickfix/createLabel/returnNoLambda.kt
@@ -1,0 +1,7 @@
+// "Create label foo@" "false"
+// ACTION: Convert to expression body
+// ERROR: Unresolved reference: @foo
+
+fun test(): Int {
+    return@<caret>foo 1
+}